refactor(firmware): MONOBASS idle grid, WF chroma auto-enable, audit fixes
- MONOBASS idle display: knob reference grid with 4 rows × 2 knob icons
  fills scope area when no note is held
- MONOBASS auto-enables WF CHROMA on entry; restores previous state on exit
- Note indicator repositioned 4 px lower for border clearance
- Audit: null check on ResponsiveAnalogRead allocation with Serial halt
- Audit: armPulseCountdown cast widened to uint32_t to prevent overflow
- Audit: extern bool monoBassKeyEvent declared in monobass.h
- Audit: memset(heldKeys, -1) replaced with range-for fill
- Audit: FM synthesis const float promoted to static constexpr
- Audit: loop counters standardized to int across DrumSynth.ino,
  monobass.h, and eeprom.h
- Audit: inline removed from 7 single-TU functions
- Audit: clarifying comments on date-parsing chain and choke range math

+++ eeprom.h
@@ -244,7 +244,7 @@ void loadPpqnFromEEPROM() {
   if (magic == EEPROM_PPQN_MAGIC) {
     uint8_t val = EEPROM.read(EEPROM_PPQN_ADDR + 1);
     // Validate: must be one of the allowed values
-    for (uint8_t i = 0; i < PPQN_OPTION_COUNT; i++) {
+    for (int i = 0; i < PPQN_OPTION_COUNT; i++) {
       if (PPQN_OPTIONS[i] == val) { ppqn = val; return; }
     }
   }
@@ -254,7 +254,7 @@ void loadPpqnFromEEPROM() {
 void savePpqnToEEPROM(uint8_t val) {
   // Only save values from the valid PPQN set
   bool valid = false;
-  for (uint8_t i = 0; i < PPQN_OPTION_COUNT; i++) {
+  for (int i = 0; i < PPQN_OPTION_COUNT; i++) {
     if (PPQN_OPTIONS[i] == val) { valid = true; break; }
   }
   if (!valid) return;
